new scripts to complete download
Had to restart the pymongo downloading script because for some reason the download speed dropped to a few KB/sec. Neither Alex nor I could figure out the reason, but restarting the script worked.

I copied both the bash and python scripts to new versions that ensured the data we had already downloaded would not have to be redownloaded.